functionality to simulate events in BIOSANS

See merge request sns-hfir-scse/sans/sans-backend!1007

Former-commit-id: f2c790ff07c2dd61bd4aed67f386cb3fe958b201

+++ drtsans/samplelogs.py
@@ -54,7 +54,7 @@ class SampleLogs(object):
         ----------
         name: str
             log entry name
-        value: str, int, double, list
+        value: str, int, float, list
             Value to insert
         unit: str
             Log unit
